Fix check mate bug
Check mate condition which was checking if any of the friendly pieces can
block the check was not working properly. In addition, I spotted that the checking squares were not calculated properly. I made some changes to the code to fix these issues. It is working so far, but I need to test it more.